chore: refactor some component names
