+++ packages/desktop/App.svelte
@@ -43,11 +43,11 @@
         Import,
         Ledger,
         Legal,
-        Login,
+        LoginRouter,
         Migrate,
         Password,
         Profile,
-        Protect,
+        ProtectRouter,
         Secure,
         Settings,
         Setup,
@@ -231,7 +231,7 @@
             <Password locale={$_} />
         </Route>
         <Route route={AppRoute.Protect} transition={false}>
-            <Protect locale={$_} />
+            <ProtectRouter />
         </Route>
         <Route route={AppRoute.Backup} transition={false}>
             <Backup locale={$_} />
@@ -255,7 +255,7 @@
             <Dashboard locale={$_} {goto} />
         </Route>
         <Route route={AppRoute.Login}>
-            <Login locale={$_} {goto} />
+            <LoginRouter {goto} />
         </Route>
         {#if settings}
             <Settings locale={$_} handleClose={() => (settings = false)} />
