+++ packages/shared/lib/core/router/subrouters/login-router.ts
@@ -8,7 +8,7 @@ export const loginRoute = writable<LoginRoute>(null)
 
 export class LoginRouter extends Subrouter<LoginRoute> {
     constructor() {
-        super(LoginRoute.Init, loginRoute)
+        super(LoginRoute.SelectProfile, loginRoute)
     }
 
     next(event?: FireflyEvent): void {
@@ -16,7 +16,7 @@ export class LoginRouter extends Subrouter<LoginRoute> {
         const currentRoute = get(this.routeStore)
 
         switch (currentRoute) {
-            case LoginRoute.Init: {
+            case LoginRoute.SelectProfile: {
                 if (event?.shouldAddProfile) {
                     get(appRouter).next(event)
                 } else {
