+++ packages/shared/routes/login/views/index.js
@@ -1,2 +1,2 @@
-export { default as EnterPin } from './EnterPin.svelte'
-export { default as SelectProfile } from './SelectProfile.svelte'
+export { default as EnterPinView } from './EnterPinView.svelte'
+export { default as SelectProfileView } from './SelectProfileView.svelte'
