+++ packages/shared/routes/setup/protect/views/ChooseProtectionMethodView.svelte
